Cloud strategy- for a retail shoe shop
I generated strategy document for modernizing the website and hosting it
to cloud. Analyzed and developed the overall enterprise solution for
retail operator in Toronto, ON. I assisted in concept design deployment
/ migration of legacy to Azure platform.This helped the business owner
to leverage on the cloud platform and move his few static web pages to
cloud and enabled him to scale and have a hi-availability for better
customer satisfaction.